requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

# LangChain and ecosystem
langchain-core
langchain-community
langchain-openai
langchain-weaviate

# LLMs and embeddings
openai==1.78.1
sentence-transformers

# Weaviate Cloud client
weaviate-client

# LangGraph (if you use it)
langgraph

# Optional: for logging and debugging
httpx
